+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/Client.java
@@ -16,6 +16,9 @@
 
 package io.fabric8.kubernetes.client;
 
+import io.fabric8.kubernetes.api.model.APIGroup;
+import io.fabric8.kubernetes.api.model.APIGroupList;
+import io.fabric8.kubernetes.api.model.APIResourceList;
 import io.fabric8.kubernetes.api.model.RootPaths;
 
 import java.io.Closeable;
@@ -48,5 +51,26 @@ public interface Client extends ConfigAware, Closeable {
    */
   boolean supportsApiPath(String path);
 
+  @Override
   void close();
+  
+  /**
+   * Returns the api groups
+   * @return the {@link APIGroupList} metadata
+   */
+  APIGroupList getApiGroups();
+  
+  /**
+   * Return a single api group
+   * @param name of the group
+   * @return the {@link APIGroup} metadata
+   */
+  APIGroup getApiGroup(String name);
+  
+  /**
+   * Return the api resource metadata for the given groupVersion
+   * @param the groupVersion - group/version
+   * @return the {@link APIResourceList} for the groupVersion
+   */
+  APIResourceList getApiResources(String groupVersion);
 }