+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/Config.java
@@ -56,6 +56,7 @@ import io.fabric8.kubernetes.client.utils.IOHelpers;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import io.fabric8.kubernetes.client.utils.Utils;
 import io.sundr.builder.annotations.Buildable;
+import okhttp3.OkHttpClient;
 import okhttp3.TlsVersion;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -1354,5 +1355,9 @@ public class Config {
 
   public AuthProviderConfig getAuthProvider() {
     return authProvider;
+  }
+
+  public OkHttpClient adaptClient(OkHttpClient httpClient) {
+    return httpClient;
   }
 }