+++ openshift-client/src/main/java/io/fabric8/openshift/client/OpenshiftAdapterSupport.java
@@ -16,36 +16,25 @@
 
 package io.fabric8.openshift.client;
 
-import io.fabric8.kubernetes.api.model.APIGroup;
-import io.fabric8.kubernetes.api.model.APIGroupList;
 import io.fabric8.kubernetes.client.BaseClient;
 import io.fabric8.kubernetes.client.Client;
 import io.fabric8.kubernetes.client.KubernetesClientException;
-import io.fabric8.kubernetes.client.utils.Serialization;
-import io.fabric8.kubernetes.client.utils.URLUtils;
 import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
 
 import java.net.URI;
-import java.net.URL;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-
-import static io.fabric8.openshift.client.DefaultOpenShiftClient.clientWithOpenShiftOAuthInterceptor;
 
 public class OpenshiftAdapterSupport {
 
-  static final ConcurrentMap<URL, Boolean> IS_OPENSHIFT = new ConcurrentHashMap<>();
-  static final ConcurrentMap<URL, Boolean> USES_OPENSHIFT_APIGROUPS = new ConcurrentHashMap<>();
-  public static final String APIS = "/apis";
-
   public Boolean isAdaptable(Client client) {
-    OpenShiftConfig config = new OpenShiftConfig(client.getConfiguration());
-    if (!hasCustomOpenShiftUrl(config) && !isOpenShift(client, config)) {
-      return false;
+    OpenShiftConfig config = OpenShiftConfig.wrap(client.getConfiguration());
+    return hasCustomOpenShiftUrl(config) || isOpenShift(client.adapt(OkHttpClient.class), config);
+  }
+
+  public DefaultOpenShiftClient adapt(Client client) {
+    if (!isAdaptable(client)) {
+      throw new OpenShiftNotAvailableException("OpenShift is not available. Root paths at: " + client.getMasterUrl() + " do not include oapi.");
     }
-    return true;
+    return new DefaultOpenShiftClient(client.adapt(OkHttpClient.class), OpenShiftConfig.wrap(client.getConfiguration()));
   }
 
   /**
@@ -54,62 +43,10 @@ public class OpenshiftAdapterSupport {
    * @param config {@link OpenShiftConfig} OpenShift Configuration
    * @return         True if oapi is found in the root paths.
    */
-  static boolean isOpenShift(Client client, OpenShiftConfig config) {
-    return isOpenShiftAPIGroups(client, config);
+  public static boolean isOpenShift(OkHttpClient client, OpenShiftConfig config) {
+    return new BaseClient(client, config) {}.getApiGroups()
+        .getGroups().stream().anyMatch(g -> g.getName().endsWith("openshift.io"));
   }
-
-  /**
-   * Check if OpenShift API Groups are available
-   * @param client   The client.
-   * @param config {@link OpenShiftConfig} OpenShift config
-   * @return         True if the new <code>/apis/*.openshift.io/</code> APIs are found in the root paths.
-   */
-  static boolean isOpenShiftAPIGroups(Client client, OpenShiftConfig config) {
-    URL masterUrl = client.getMasterUrl();
-
-    OkHttpClient httpClient = clientWithOpenShiftOAuthInterceptor(((BaseClient)client).getHttpClient(), config);
-    try {
-      Request.Builder requestBuilder = new Request.Builder()
-        .get()
-        .url(URLUtils.join(masterUrl.toString(), APIS));
-      Response response = httpClient.newCall(requestBuilder.build()).execute();
-      APIGroupList apiGroupList = Serialization.unmarshal(response.body().string(), APIGroupList.class);
-
-      for (APIGroup apiGroup : apiGroupList.getGroups()) {
-        if (apiGroup.getName().endsWith("openshift.io")) {
-          return true;
-        }
-      }
-    } catch(Exception e) {
-      KubernetesClientException.launderThrowable(e);
-    }
-    return false;
-  }
-
-    /**
-     * Check if OpenShift API Groups are available
-     * @param httpClient   The httpClient.
-     * @param masterUrl    The master url.
-     * @return             True if the new <code>/apis/*.openshift.io/</code> APIs are found in the root paths.
-     */
-    static boolean isOpenShiftAPIGroups(OkHttpClient httpClient, String masterUrl) {
-      try {
-        Request.Builder requestBuilder = new Request.Builder()
-          .get()
-          .url(URLUtils.join(masterUrl, APIS));
-        Response response = httpClient.newCall(requestBuilder.build()).execute();
-        APIGroupList apiGroupList = Serialization.unmarshal(response.body().string(), APIGroupList.class);
-
-        for (APIGroup apiGroup : apiGroupList.getGroups()) {
-          if (apiGroup.getName().endsWith("openshift.io")) {
-            return true;
-          }
-        }
-      } catch(Exception e) {
-        KubernetesClientException.launderThrowable(e);
-      }
-      return false;
-    }
 
     /**
      * Checks if a custom URL for OpenShift has been used.